Move keyboard input settling to ConvosCoreiOS to fix archive type-check timeout
The Dev archive fails with "expression took 610ms to type-check (limit:
300ms)" at FocusCoordinator.swift:252. The cost is not solver ambiguity:
the first member lookup through the (UIResponder & UITextInput)
existential in the app module pays a lazy ObjC member-table load that
takes 600ms+ given the app target's import graph, and the per-expression
timer charges it all to whichever expression touches it first. Typed
hoists and narrowing the existential just move the charge to the next
lookup (verified empirically: 670ms, 640ms, then 614ms on
resignFirstResponder after removing the first lookup).

Fix by moving the settle/handoff machinery verbatim into a new
KeyboardInputSettling helper in ConvosCoreiOS, where SPM targets compile
without the timing flags and the lookup is also genuinely cheap.
FocusCoordinator.withSettledKeyboardInput keeps its API and delegates,
so the app target no longer contains any member access through that
existential. Behavior is unchanged.
